Implemented security rule for public ip on emulated cloud

+++ src/main/java/cloud/fogbow/ras/core/plugins/interoperability/emulatedcloud/securityrule/EmulatedCloudSecurityRulePlugin.java
@@ -9,11 +9,12 @@ import cloud.fogbow.ras.api.parameters.SecurityRule;
 import cloud.fogbow.ras.core.models.orders.Order;
 import cloud.fogbow.ras.core.plugins.interoperability.SecurityRulePlugin;
 import cloud.fogbow.ras.core.plugins.interoperability.emulatedcloud.EmulatedCloudUtils;
+import cloud.fogbow.ras.core.plugins.interoperability.emulatedcloud.emulatedmodels.EmulatedOrderWithSecurityRule;
 import cloud.fogbow.ras.core.plugins.interoperability.emulatedcloud.emulatedmodels.EmulatedPublicIp;
 import cloud.fogbow.ras.core.plugins.interoperability.emulatedcloud.emulatedmodels.EmulatedSecurityRule;
+import cloud.fogbow.ras.core.plugins.interoperability.emulatedcloud.emulatedmodels.EmulatedSecurityRuleBinding;
 
 import java.io.IOException;
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Properties;
 
@@ -30,21 +31,45 @@ public class EmulatedCloudSecurityRulePlugin implements SecurityRulePlugin<Cloud
 
         EmulatedSecurityRule emulatedSecurityRule = generateJsonEntityToCreateInstance(String);
 
-        String instanceId = majorOrder.getInstanceId();
-        this.addSecurityRuleToOrder(instanceId, emulatedSecurityRule);
+        String majorOrderInstanceId = majorOrder.getInstanceId();
+        this.addSecurityRuleToOrder(majorOrderInstanceId, emulatedSecurityRule);
 
         String securityRuleInstanceId = emulatedSecurityRule.getId();
+        this.bindSecurityRuleToOrder(securityRuleInstanceId, majorOrderInstanceId);
+
         return securityRuleInstanceId;
     }
 
     @Override
     public List<SecurityRuleInstance> getSecurityRules(Order majorOrder, CloudUser cloudUser) throws FogbowException {
-        return new ArrayList<>();
+        String instanceId = majorOrder.getInstanceId();
+
+        List<SecurityRuleInstance> securityRuleInstances = null;
+        try {
+
+            EmulatedPublicIp publicIp = this.getPublicIpById(instanceId);
+            List<EmulatedSecurityRule> securityRules = publicIp.getSecurityRules();
+            securityRuleInstances = EmulatedOrderWithSecurityRule.getFogbowSecurityRules(securityRules);
+
+        } catch (FogbowException e) {
+
+            throw new FogbowException(e.getMessage());
+        }
+
+        return securityRuleInstances;
     }
 
     @Override
-    public void deleteSecurityRule(String StringId, CloudUser cloudUser) throws FogbowException {
+    public void deleteSecurityRule(String securityRuleId, CloudUser cloudUser) throws FogbowException {
+        String majorOrderId = getMajorOrderIdFromSecurityRule(securityRuleId);
+        this.removeSecurityRuleFromOrder(majorOrderId, securityRuleId);
 
+        String securityRulePath = EmulatedCloudUtils.getResourcePath(this.properties, securityRuleId);
+        try {
+            EmulatedCloudUtils.deleteFile(securityRulePath);
+        } catch (Exception e) {
+            throw new FogbowException(e.getMessage());
+        }
     }
 
     private EmulatedSecurityRule generateJsonEntityToCreateInstance(SecurityRule String) {
@@ -83,10 +108,18 @@ public class EmulatedCloudSecurityRulePlugin implements SecurityRulePlugin<Cloud
     }
 
     private void updateResourceOnDisk(String instanceId, JsonSerializable serializableOrder) throws IOException {
+        updateResourceOnDisk(instanceId, serializableOrder, true);
+    }
+
+    private void updateResourceOnDisk(String instanceId, JsonSerializable serializableOrder, boolean deletePrevious) throws IOException {
         String contents = serializableOrder.toJson();
         String resourcePath = EmulatedCloudUtils.getResourcePath(this.properties, instanceId);
 
-        EmulatedCloudUtils.deleteFile(resourcePath);
+        if (deletePrevious) {
+
+            EmulatedCloudUtils.deleteFile(resourcePath);
+        }
+
         EmulatedCloudUtils.saveFileContent(resourcePath, contents);
     }
 
@@ -99,5 +132,51 @@ public class EmulatedCloudSecurityRulePlugin implements SecurityRulePlugin<Cloud
         } catch (IOException e) {
             throw new FogbowException(e.getMessage());
         }
+    }
+
+    private void removeSecurityRuleFromOrder(String majorOrderId, String securityRuleId) throws FogbowException {
+        try {
+            EmulatedPublicIp publicIp = this.getPublicIpById(majorOrderId);
+            publicIp.removeSecurityRule(securityRuleId);
+
+        } catch (Exception e) {
+            throw new FogbowException(e.getMessage());
+        }
+    }
+
+    private void bindSecurityRuleToOrder(String securityRuleInstanceId, String majorOrderInstanceId) throws FogbowException {
+        EmulatedSecurityRuleBinding securityRuleBinding = new EmulatedSecurityRuleBinding.Builder()
+                .majorOrderId(majorOrderInstanceId)
+                .build();
+
+        try {
+            updateResourceOnDisk(securityRuleInstanceId, securityRuleBinding);
+        } catch (IOException e) {
+            throw new FogbowException(e.getMessage());
+        }
+    }
+
+    private EmulatedSecurityRuleBinding getEmulatedSecurityRuleBindingById(String securityRuleId) throws FogbowException {
+        String securityRulePath = EmulatedCloudUtils.getResourcePath(this.properties, securityRuleId);
+
+        String securityRuleJson = null;
+        EmulatedSecurityRuleBinding emulatedSecurityRuleBinding = null;
+
+        try {
+
+            securityRuleJson = EmulatedCloudUtils.getFileContent(securityRulePath);
+            emulatedSecurityRuleBinding = EmulatedSecurityRuleBinding.fromJson(securityRuleJson);
+        } catch (IOException e) {
+
+            throw new FogbowException(e.getMessage());
+        }
+
+        return emulatedSecurityRuleBinding;
+    }
+
+    private String getMajorOrderIdFromSecurityRule(String securityRuleId) throws FogbowException {
+        EmulatedSecurityRuleBinding emulatedSecurityRuleBinding = getEmulatedSecurityRuleBindingById(securityRuleId);
+
+        return emulatedSecurityRuleBinding.getMajorOrderId();
     }
 }
